moved pvp stuff
