+++ frontend/src/components/layout/navigation/PvPNavBar.vue
