+++ frontend/src/components/pvp/PvPArenaPreparation.vue
@@ -198,9 +198,9 @@
 import { mapState, mapActions } from 'vuex';
 import BN from 'bignumber.js';
 import { BPopover } from 'bootstrap-vue';
-import PvPWeapon from './PvPWeapon.vue';
-import PvPShield from './PvPShield.vue';
-import PvPCharacter from './PvPCharacter.vue';
+import PvPWeapon from '../pvp/PvPWeapon.vue';
+import PvPShield from '../pvp/PvPShield.vue';
+import PvPCharacter from '../pvp/PvPCharacter.vue';
 import PvPSeparator from './PvPSeparator.vue';
 import checkIcon from '../../assets/checkImage.svg';
 import ellipseIcon from '../../assets/ellipseImage.svg';
