+++ frontend/src/components/pvp/PvPArenaSummary.vue
@@ -90,10 +90,10 @@
 <script>
 import BN from 'bignumber.js';
 import { mapState, mapActions } from 'vuex';
-import PvPWeapon from './PvPWeapon.vue';
-import PvPShield from './PvPShield.vue';
+import PvPWeapon from '../pvp/PvPWeapon.vue';
+import PvPShield from '../pvp/PvPShield.vue';
 import PvPButton from './PvPButton.vue';
-import PvPCharacter from './PvPCharacter.vue';
+import PvPCharacter from '../pvp/PvPCharacter.vue';
 import dayjs from 'dayjs';
 import PvPArenaInfo from '../pvp/PvPArenaInfo.vue';
 import i18n from '../../i18n';
