+++ frontend/src/components/pvp/PvPKickedModal.vue
