+++ frontend/src/components/pvp/PvPWeapon.vue
